requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
altair==6.0.0
annotated-types==0.7.0
anyio==3.7.1
attrs==25.4.0
blinker==1.9.0
cachetools==6.2.4
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
distro==1.9.0
frozenlist==1.8.0
gitdb==4.0.12
GitPython==3.1.46
google-ai-generativelanguage==0.6.10
google-api-core==2.28.1
google-api-python-client==2.187.0
google-auth==2.46.0
google-auth-httplib2==0.3.0
google-generativeai==0.8.3
googleapis-common-protos==1.72.0
grpcio==1.76.0
grpcio-status==1.71.2
h11==0.16.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
idna==3.11
Jinja2==3.1.6
jiter==0.12.0
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
MarkupSafe==3.0.3
multidict==6.7.0
narwhals==2.15.0
numpy==2.4.0
openai==0.28.0
packaging==25.0
pandas==2.3.3
pillow==12.1.0
pip==25.3
propcache==0.4.1
proto-plus==1.27.0
protobuf==5.29.5
pyarrow==22.0.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.12.5
pydantic_core==2.41.5
pydeck==0.9.1
pyparsing==3.3.1
python-dateutil==2.9.0.post0
pytz==2025.2
referencing==0.37.0
requests==2.32.5
rpds-py==0.30.0
rsa==4.9.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.52.2
tenacity==9.1.2
toml==0.10.2
tornado==6.5.4
tqdm==4.67.1
typing_extensions==4.15.0
typing-inspection==0.4.2
tzdata==2025.3
uritemplate==4.2.0
urllib3==2.6.2
watchdog==6.0.0
yarl==1.22.0
plotly==5.24.1
python-dotenv